+++ CHANGES.md
@@ -16,6 +16,7 @@ Core Changes:
 
 Language Improvements:
 
+- enh(fortran) Support Fortran 77 style comments (#2416) [Josh Goebel][]
 - (csharp) add support for `@identifier` style identifiers (#2414) [Josh Goebel][]
 - fix(elixir) Support function names with a slash (#2406) [Josh Goebel][]
 - fix(javascript) comma is allowed in a "value container" (#2403) [Josh Goebel][]